Add support for TOTP in exports
